Working configuration for fb login. At the moment works only in production. Small changes to Bust and Mark vote count.

+++ src/vahdin/data/Mark.java
@@ -66,8 +66,8 @@ public class Mark implements Item {
         return getId();
     }
 
-    public float getVoteCount() {
-        return this.voteCount; // TODO:
+    public int getVoteCount() {
+        return (int) this.voteCount; // TODO:
     }
 
     public int getId() {
